requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
dataclasses-json==0.6.7
distro==1.9.0
docstring_parser==0.16
faiss-cpu==1.7.4
Flask==3.1.0
frozenlist==1.5.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
Hypercorn==0.17.3
hyperframe==6.1.0
idna==3.10
instructor==1.7.9
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.23
langchain-community==0.3.21
langchain-core==0.3.51
# langchain-openai==0.3.12  # Replaced with DeepSeek and HuggingFace embeddings
# Using compatible versions: sentence-transformers 2.2.2 works with huggingface_hub <0.20.0
sentence-transformers==2.2.2
huggingface_hub<0.20.0,>=0.16.0
langchain-text-splitters==0.3.8
langsmith==0.3.24
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mcp==1.5.0
mdurl==0.1.2
multidict==6.3.2
mypy-extensions==1.0.0
mysql-connector-python==9.2.0
numpy==1.26.4
openai==1.70.0
orjson==3.10.16
packaging==23.2
pandas==2.2.3
priority==2.0.0
propcache==0.3.1
pydantic==2.11.2
pydantic-settings==2.8.1
pydantic_core==2.33.1
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
Quart==0.20.0
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
sse-starlette==2.2.1
starlette==0.46.1
tenacity==9.1.2
tiktoken==0.9.0
tqdm==4.67.1
typer==0.15.2
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2025.2
urllib3==2.3.0
uvicorn==0.34.0
Werkzeug==3.1.3
wsproto==1.2.0
yarl==1.19.0
zstandard==0.23.0
